src/requirements.txt
aiohttp==3.9.5
aiohttp-compress==0.2.1
azure-core==1.31.0
azure-identity==1.19.0
azure-cosmos==3.2.0
azure-search==1.0.0b2
azure-search-documents==11.5.1
botbuilder-azure==4.16.1
botbuilder-core==4.16.1
botbuilder-dialogs==4.16.1
botbuilder-integration-aiohttp==4.16.1
python-dotenv==1.0.1
openai==1.41.0
semantic-kernel==1.9.0